requirements.txt
altair
numpy
pandas
pydeck
streamlit
python-pptx
matplotlib
seaborn